feat(core): add model compted query

+++ packages/core/__test__/prisma.ts
@@ -1,14 +1,23 @@
 import client from '@prisma/client'
 
 if (!client.PrismaClient) {
-  throw new Error('[setPrisma] error, prisma.PrismaClient not found please run prisma generate first')
+  throw new Error(
+    '[setPrisma] error, prisma.PrismaClient not found please run prisma generate first'
+  )
 }
 
 const prisma = new client.PrismaClient()
 prisma.$connect()
 
-export async function clearAll () {
-  prisma.item.deleteMany({})
+export async function clearAll() {
+  try {
+    const all = await prisma.item.findMany({ where: {} })
+    if (all.length > 0) {
+      await prisma.item.deleteMany({})
+    }
+  } catch(e) {
+    console.error('clearAll', e)
+  }
 }
 
 export default prisma
